all-python-files/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
PyJWT==2.9.0
bcrypt==4.2.1
python-multipart==0.0.12
pydantic==2.9.2
aiofiles==24.1.0
httpx==0.27.2
